+++ packages/global.d.ts
@@ -1,0 +1,3 @@
+declare type Result<T> =
+  | { ok: true }
+  | { ok: false; error: T };